debuggin galore

+++ ulmo/runs/SSL/MODIS/v2/ssl_modis_v2.py
@@ -43,7 +43,7 @@ def parse_option():
     
     return args
 
-def ssl_v2_umap(debug=False, orig=False):
+def ssl_v2_umap(debug=False):
     """Run a UMAP analysis on all the MODIS L2 data
 
     Args:
@@ -63,12 +63,10 @@ def ssl_v2_umap(debug=False, orig=False):
     valid = modis_tbl.pp_type == 0
     y2010 = modis_tbl.pp_file == 's3://modis-l2/PreProc/MODIS_R2019_2010_95clear_128x128_preproc_std.h5'
     valid_tbl = modis_tbl[valid & y2010].copy()
+    nvalid = len(valid_tbl)
 
     # Latents file (subject to move)
-    if debug:
-        latents_train_file = 's3://modis-l2/SSL_MODIS_R2019_2010_latents_v2/modis_R2019_2010_latents_last_v2.h5'
-    else:
-        latents_train_file = 's3://modis-l2/SSL/SSL_v2_2012/latents/MODIS_R2019_2010_95clear_128x128_latents_std.h5'
+    latents_train_file = 's3://modis-l2/SSL/SSL_v2_2012/latents/MODIS_R2019_2010_95clear_128x128_latents_std.h5'
 
     # Load em in
     basefile = os.path.basename(latents_train_file)
@@ -77,14 +75,19 @@ def ssl_v2_umap(debug=False, orig=False):
         ulmo_io.download_file_from_s3(basefile, latents_train_file)
         print("Done")
     hf = h5py.File(basefile, 'r')
-    latents_train = hf['modis_latents_v2_train'][:]
-    latents_valid = hf['modis_latents_v2_valid'][:]
+    latents_valid = hf['valid'][:]
     print("Latents loaded")
 
     # Check
-    assert latents_valid.shape[0] == len(valid_tbl)
+    assert latents_valid.shape[0] == nvalid
 
-    print("Running UMAP..")
+    ntrain = 150000
+    train_idx = np.arange(nvalid)
+    np.random.shuffle(train_idx)
+    train_idx = train_idx[0:ntrain]
+    latents_train = latents_valid[train_idx]
+
+    print(f"Running UMAP on a {ntrain} subset of valid..")
     reducer_umap = umap.UMAP()
     latents_mapping = reducer_umap.fit(latents_train)
     print("Done..")
@@ -92,6 +95,8 @@ def ssl_v2_umap(debug=False, orig=False):
     # Loop on em all
     latent_files = ulmo_io.list_of_bucket_files('modis-l2',
                                                 prefix='SSL/SSL_v2_2012/latents/')
+    if debug:
+        latent_files = latent_files[0:1]
 
     for latents_file in latent_files:
         basefile = os.path.basename(latents_file)
@@ -104,29 +109,28 @@ def ssl_v2_umap(debug=False, orig=False):
 
         #  Load and apply
         hf = h5py.File(basefile, 'r')
-        '''
+
+        # Train
         if 'train' in hf.keys():
             latents_train = hf['train'][:]
             train_embedding = latents_mapping.transform(latents_train)
-        '''
 
-        # THIS LINE IS WRONG.  FIX WHEN THE REST IS FIXED
-        latents_valid = hf['train'][:]
+        # Valid
+        latents_valid = hf['valid'][:]
         valid_embedding = latents_mapping.transform(latents_valid)
 
         # Save to table
-        embed(header='118 of ssl modis 2012')
         yidx = modis_tbl.pp_file == f's3://modis-l2/PreProc/MODIS_R2019_{year}_95clear_128x128_preproc_std.h5'
         valid_idx = valid & yidx
         modis_tbl.loc[valid_idx, 'U0'] = valid_embedding[:,0]
         modis_tbl.loc[valid_idx, 'U1'] = valid_embedding[:,1]
-
-        '''
+        
+        # Train?
         train_idx = train & yidx
-        if np.sum(train_idx) > 0:
+        if 'train' in hf.keys() and (np.sum(train_idx) > 0):
             modis_tbl.loc[train_idx, 'U0'] = train_embedding[:,0]
             modis_tbl.loc[train_idx, 'U1'] = train_embedding[:,1]
-        '''
+
 
         hf.close()
 
@@ -134,7 +138,11 @@ def ssl_v2_umap(debug=False, orig=False):
         os.remove(basefile)
 
     # Vet
-    assert cat_utils.vet_main_table(valid_tbl, cut_prefix='modis_')
+    assert cat_utils.vet_main_table(modis_tbl, cut_prefix='modis_')
+
+    # Final write
+    if not debug:
+        ulmo_io.write_main_table(modis_tbl, tbl_file) 
 
 
 def main_train(opt_path: str):
@@ -213,6 +221,7 @@ def main_evaluate(opt_path, model_file,
     key_train, key_valid = "train", "valid"
     if debug:
         pp_files = pp_files[0:1]
+
     for ifile in pp_files:
         print("Working on ifile")
         data_file = os.path.basename(ifile)
@@ -245,8 +254,9 @@ def main_evaluate(opt_path, model_file,
         ulmo_io.upload_file_to_s3(latents_file, latents_path)
 
         # Remove data file
-        os.remove(data_file)
-        print(f'{data_file} removed')
+        if not debug:
+            os.remove(data_file)
+            print(f'{data_file} removed')
         
 if __name__ == "__main__":
     # get the argument of training.
@@ -269,5 +279,7 @@ if __name__ == "__main__":
     # run the umap
     if args.func_flag == 'umap':
         print("UMAP Starts.")
+        if args.debug:
+            print("In debug mode!!")
         ssl_v2_umap(debug=args.debug)
         print("UMAP Ends.")
